2025-11-21-guardrails_of_fitness/backend/pyproject.toml
[project]
name = "fitness-analysis-backend"
version = "0.1.0"
description = "Fitness analysis backend with LangGraph and Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.1",
    "langgraph>=0.2.60",
    "langchain-google-genai>=2.0.5",
    "langchain-core>=0.3.28",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
]
